pyproject.toml
[tool.poetry]
name = "2geda test project"
version = "0.1.0"
description = ""
authors = ["dakohhh <wiizzydreadmill@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
pyjwt = "^2.8.0"
pydantic = "^2.6.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
